Separate core and editor types
