+++ src/cli.ts
@@ -3,8 +3,8 @@
 import fs from "node:fs";
 import { parseArgs } from "node:util";
 import { faceToSvgString } from "./faceToSvgString.js";
-import { type Gender, type Race, generate } from "./generate.js";
-import type { Overrides } from "./override.js";
+import { generate } from "./generate.js";
+import { Gender, Overrides, Race } from "./types.js";
 
 const { values: options } = parseArgs({
   options: {
